requirements.txt
ollama
pyyaml
numpy
pandas
matplotlib
seaborn
jupyter
requests
tqdm
scikit-learn
